requirements.txt
# 基础依赖
python-docx>=0.8.11  # 用于处理Word文档
pandas>=1.5.0        # 用于处理表格数据
numpy>=1.21.0        # 用于数学计算
PyPDF2>=3.0.0        # 用于处理PDF文件
python-magic>=0.4.27 # 用于文件类型检测

# 文本处理
nltk>=3.8.1         # 自然语言处理工具
spacy>=3.5.0        # 高级NLP处理
regex>=2022.10.31   # 增强的正则表达式支持

# 表格处理
tabula-py>=2.6.0    # 用于从PDF提取表格
openpyxl>=3.1.0     # 用于处理Excel文件

# 公式处理
sympy>=1.11.0       # 用于数学公式处理
latex2mathml>=3.75.0 # 用于LaTeX公式转换

# 工具和辅助
tqdm>=4.65.0        # 进度条显示
pydantic>=2.0.0     # 数据验证
python-dotenv>=1.0.0 # 环境变量管理

openai>=1.0.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn==1.3.0
tensorflow==2.13.0
fastapi==0.101.1
uvicorn==0.23.2
python-multipart==0.0.6
sentencepiece==0.1.99
protobuf==3.20.3 